refactoring: Extract abstract class
