+++ src/main/java/it/gov/pagopa/wispconverter/servicebus/PaymentTimeoutConsumer.java
@@ -5,7 +5,6 @@ import com.azure.messaging.servicebus.*;
 import java.io.IOException;
 import java.time.ZonedDateTime;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
 
@@ -22,7 +21,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 @Slf4j
-public class PaymentTimeoutConsumer {
+public class PaymentTimeoutConsumer extends SBConsumer {
 
     @Value("${azure.sb.connectionString}")
     private String connectionString;
@@ -76,37 +75,6 @@ public class PaymentTimeoutConsumer {
         } catch (IOException e) {
             log.error("Error when read ReceiptDto value from message: '{}'. Body: '{}'",
                     message.getMessageId(), message.getBody());
-        }
-    }
-
-    public void processError(ServiceBusErrorContext context) {
-        log.error("Error when receiving messages from namespace: '{}'. Entity: '{}'",
-                context.getFullyQualifiedNamespace(), context.getEntityPath());
-
-        if (!(context.getException() instanceof ServiceBusException exception)) {
-            log.error("Non-ServiceBusException occurred: {}", context.getException().getMessage());
-            return;
-        }
-
-        ServiceBusFailureReason reason = exception.getReason();
-
-        if (reason == ServiceBusFailureReason.MESSAGING_ENTITY_DISABLED
-                || reason == ServiceBusFailureReason.MESSAGING_ENTITY_NOT_FOUND
-                || reason == ServiceBusFailureReason.UNAUTHORIZED) {
-            log.error("An unrecoverable error occurred. Stopping processing with reason {}: {}",
-                    reason, exception.getMessage());
-        } else if (reason == ServiceBusFailureReason.MESSAGE_LOCK_LOST) {
-            log.error("Message lock lost for message: {}", context.getException().getMessage());
-        } else if (reason == ServiceBusFailureReason.SERVICE_BUSY) {
-            try {
-                // Choosing an arbitrary amount of time to wait until trying again.
-                TimeUnit.SECONDS.sleep(1);
-            } catch (InterruptedException e) {
-                log.debug("Unable to sleep for period of time");
-            }
-        } else {
-            log.error("Error source {}, reason {}, message: {}", context.getErrorSource(),
-                    reason, context.getException());
         }
     }
 }
